requirements.txt
numpy==1.25
chatterbox-tts
torch
flask
flask-cors
